v33: tech debt cleanup — real health check, CID index, MD chunking, file import, portability
- Fix B18 test: use total() instead of summing 6/7 weights
- Fix oldest_entry_age_ms: track max age (oldest) not min age (newest)
- llm_available(): real probe request instead of stub returning true
- is_cid_referenced(): O(1) reverse index on PetgraphBackend (was O(n) scan)
- Markdown-aware chunking: split on headings/thematic breaks, preserve code blocks
- CLI file import: aicli put --file/--dir with recursive glob + auto-chunking
- API file import: ApiRequest::ImportFiles with per-file results
- Unified PlicoConfig: 3-layer cascade (defaults → config.json → env vars)
- Portability: cfg(unix) for UDS, temp_dir(), cross-platform memory stats
- Clippy: 15 warnings → 0; ApiResponse Default impl
- tenant.rs TODOs → ROADMAP (deferred to multi-tenant milestone)
- Tests: 1,038 unit + 38 integration — all green

Made-with: Cursor

+++ src/memory/pressure.rs
@@ -43,7 +43,7 @@ pub fn eviction_priority(entry: &MemoryEntry, now_ms: u64) -> u32 {
         5
     };
 
-    let access_bonus = (entry.access_count.min(50) as u32) * 2;
+    let access_bonus = entry.access_count.min(50) * 2;
 
     let superseded_penalty = if entry.supersedes.is_some() || is_likely_superseded(entry) {
         0
